+++ bindings/python/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.11.2] - 2023-12-09
+
 ### Performance
 
 - Avoid iterating over non-Element nodes.
@@ -363,7 +365,8 @@
 
 - Initial public release
 
-[Unreleased]: https://github.com/Stranger6667/css-inline/compare/python-v0.11.1...HEAD
+[Unreleased]: https://github.com/Stranger6667/css-inline/compare/python-v0.11.2...HEAD
+[0.11.2]: https://github.com/Stranger6667/css-inline/compare/python-v0.11.1...python-v0.11.2
 [0.11.1]: https://github.com/Stranger6667/css-inline/compare/python-v0.11.0...python-v0.11.1
 [0.11.0]: https://github.com/Stranger6667/css-inline/compare/python-v0.10.5...python-v0.11.0
 [0.10.5]: https://github.com/Stranger6667/css-inline/compare/python-v0.10.4...python-v0.10.5